create statement object
