+++ go/adbc/driver/snowflake/statement.go
@@ -1,0 +1,148 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package snowflake
+
+import (
+	"context"
+
+	"github.com/apache/arrow-adbc/go/adbc"
+	"github.com/apache/arrow/go/v12/arrow"
+	"github.com/apache/arrow/go/v12/arrow/array"
+	"github.com/apache/arrow/go/v12/arrow/memory"
+)
+
+type statement struct {
+	cnxn  *cnxn
+	alloc memory.Allocator
+}
+
+// Close releases any relevant resources associated with this statement
+// and closes it (particularly if it is a prepared statement).
+//
+// A statement instance should not be used after Close is called.
+func (st *statement) Close() error {
+	// noop
+	return nil
+}
+
+// SetOption sets a string option on this statement
+func (st *statement) SetOption(key string, val string) error {
+	return adbc.Error{
+		Msg:  "SetOption not implemented for Snowflake.Statement",
+		Code: adbc.StatusNotImplemented,
+	}
+}
+
+// SetSqlQuery sets the query string to be executed.
+//
+// The query can then be executed with any of the Execute methods.
+// For queries expected to be executed repeatedly, Prepare should be
+// called before execution.
+func (st *statement) SetSqlQuery(query string) error {
+	panic("not implemented") // TODO: Implement
+}
+
+// ExecuteQuery executes the current query or prepared statement
+// and returnes a RecordReader for the results along with the number
+// of rows affected if known, otherwise it will be -1.
+//
+// This invalidates any prior result sets on this statement.
+func (st *statement) ExecuteQuery(_ context.Context) (array.RecordReader, int64, error) {
+	panic("not implemented") // TODO: Implement
+}
+
+// ExecuteUpdate executes a statement that does not generate a result
+// set. It returns the number of rows affected if known, otherwise -1.
+func (st *statement) ExecuteUpdate(_ context.Context) (int64, error) {
+	panic("not implemented") // TODO: Implement
+}
+
+// Prepare turns this statement into a prepared statement to be executed
+// multiple times. This invalidates any prior result sets.
+func (st *statement) Prepare(_ context.Context) error {
+	panic("not implemented") // TODO: Implement
+}
+
+// SetSubstraitPlan allows setting a serialized Substrait execution
+// plan into the query or for querying Substrait-related metadata.
+//
+// Drivers are not required to support both SQL and Substrait semantics.
+// If they do, it may be via converting between representations internally.
+//
+// Like SetSqlQuery, after this is called the query can be executed
+// using any of the Execute methods. If the query is expected to be
+// executed repeatedly, Prepare should be called first on the statement.
+func (st *statement) SetSubstraitPlan(plan []byte) error {
+	return adbc.Error{
+		Msg:  "Snowflake does not support Substrait plans",
+		Code: adbc.StatusNotImplemented,
+	}
+}
+
+// Bind uses an arrow record batch to bind parameters to the query.
+//
+// This can be used for bulk inserts or for prepared statements.
+// The driver will call release on the passed in Record when it is done,
+// but it may not do this until the statement is closed or another
+// record is bound.
+func (st *statement) Bind(ctx context.Context, values arrow.Record) error {
+	panic("not implemented") // TODO: Implement
+}
+
+// BindStream uses a record batch stream to bind parameters for this
+// query. This can be used for bulk inserts or prepared statements.
+//
+// The driver will call Release on the record reader, but may not do this
+// until Close is called.
+func (st *statement) BindStream(ctx context.Context, stream array.RecordReader) error {
+	panic("not implemented") // TODO: Implement
+}
+
+// GetParameterSchema returns an Arrow schema representation of
+// the expected parameters to be bound.
+//
+// This retrieves an Arrow Schema describing the number, names, and
+// types of the parameters in a parameterized statement. The fields
+// of the schema should be in order of the ordinal position of the
+// parameters; named parameters should appear only once.
+//
+// If the parameter does not have a name, or a name cannot be determined,
+// the name of the corresponding field in the schema will be an empty
+// string. If the type cannot be determined, the type of the corresponding
+// field will be NA (NullType).
+//
+// This should be called only after calling Prepare.
+//
+// This should return an error with StatusNotImplemented if the schema
+// cannot be determined.
+func (st *statement) GetParameterSchema() (*arrow.Schema, error) {
+	panic("not implemented") // TODO: Implement
+}
+
+// ExecutePartitions executes the current statement and gets the results
+// as a partitioned result set.
+//
+// It returns the Schema of the result set, the collection of partition
+// descriptors and the number of rows affected, if known. If unknown,
+// the number of rows affected will be -1.
+//
+// If the driver does not support partitioned results, this will return
+// an error with a StatusNotImplemented code.
+func (st *statement) ExecutePartitions(_ context.Context) (*arrow.Schema, adbc.Partitions, int64, error) {
+	panic("not implemented") // TODO: Implement
+}
